+++ CHANGELOG.md
@@ -1,8 +1,16 @@
+# 0.0.3
+
+* fix static-analysis warning of calling `setState` outside a subclass of `flutter/widgets.dart`.
+* new static method `dispose`.
+* refactoring method `update`.
+* removed unnecessary generic type notation from `Hydro` abstract class.
+* updated `/example`.
+* added warning log messages in debug mode.
+
 ## 0.0.2
 
 * skip `states` cleaning when refreshing the UI.
 * fix some linting issues.
-
 
 ## 0.0.1-beta1